+++ jpo-conflictvisualizer-api/src/main/java/us/dot/its/jpo/ode/api/accessors/spat/ProcessedSpatRepositoryImpl.java
@@ -83,7 +83,8 @@ public class ProcessedSpatRepositoryImpl implements ProcessedSpatRepository {
         AggregationResults<IDCount> result = mongoTemplate.aggregate(aggregation, collectionName, IDCount.class);
         List<IDCount> results = result.getMappedResults();
         for (IDCount r: results){
-            r.setCount((int)((float)r.getCount() / 3600.0));
+            r.setCount((float)r.getCount() / 3600.0);
+            
         }
 
         return results;
